+++ R/chisq.R
@@ -2,7 +2,6 @@
 ##   Functions related to the Chi Square distribution    ##
 ## --##--##--##--##--##--##--##--##--##--##--##--##--##--##
 
-#' @importFrom stats dchisq pchisq rchisq
 #' @param df degrees of freedom for "parent" distribution
 #' @rdname rtrunc
 #' @export
@@ -38,23 +37,24 @@ sufficientT.trunc_chisq <- function(y) {
 }
 
 #' @export
-natural2parameters.trunc_chisq <- function(eta) {
+natural2parameters.trunc_chisq <- function(eta, ...) {
   # eta: The natural parameters in a Chi Square distribution
   # returns df
-  df <- c(df = 2 * (eta + 1))
+  if (length(eta) != 1) stop("Eta must be one single number")
+  df <- c(df = 2 * (eta[[1]] + 1))
   class(df) <- class(eta)
   return(df)
 }
 
 #' @export
-parameters2natural.trunc_chisq <- function(parms) {
+parameters2natural.trunc_chisq <- function(parms, ...) {
   # parms: The parameter lambda in a Chi Square distribution
   # returns the natural parameters
   eta <- prepEta(parms / 2 - 1, class(parms))
   return(eta)
 }
 
-getGradETinv.trunc_chisq <- function(eta) {
+getGradETinv.trunc_chisq <- function(eta, ...) {
   # eta: Natural parameter
   # return the inverse of E.T differentiated with respect to eta
   return(A = 1 / sum(1 / (as.vector(eta) + (1:1e6))^2))