+++ R/mlEstimationTruncDist.R
@@ -20,9 +20,6 @@
 #' maximum likelihood estimation of the parameters of the truncated normal and
 #' lognormal distributions"
 #' @author René Holst
-#' @importFrom stats dbinom dgamma dlnorm dnorm dpois pbinom pgamma plnorm pnorm
-#' ppois rbinom rgamma rlnorm rnorm rpois var
-#' @importFrom methods is
 #' @examples
 #' sample_size <- 1000
 #' # Normal
@@ -157,9 +154,6 @@ welcomeToFamily <- function(y, family) {
   if (!is.null(family)) {
     # Adding proper family attributes
     family <- useStandardFamilyName(family)
-    if (!is(y, "numeric")) {
-      message("Data is originally ", class(y), ". Treating as ", family)
-    }
     class(y) <- paste0("trunc_", family)
     attr(y, "continuous") <- valid_fam_parm[[family]][["cont"]]
     attr(y, "parameters") <- switch(